Crate schnorr

Formally verified implementation of the Schnorr protocol over elliptic curves. See theory for the properties that this implementation has been verified to satisfy.

Modules

abstract_structures

Defines traits for the underlying field of an elliptic curve and the scalar type used for scalar multiplication.

elliptic_curve

Implementation of an elliptic curve over a finite field.

point

Defines affine coordinate points

schnorr

Implementation of the Schnorr Identification Scheme over elliptic curves.

theory

Specifies the the properties of elliptic curves and the Schnorr protocol, that this implementation has been formally verified to satisfy.