schnorr

Module schnorr

Implementation of the Schnorr Identification Scheme over elliptic curves.

Functions

challenge

Bob chooses a challenge c uniformly at random from [0, 2^t-1], where t is the bit length of the challenge (say, t = 80). Bob sends c to Alice.

commitment

Alice chooses a number v uniformly at random from [1, n-1] and computes V = G x [v]. She sends V to Bob.

gen_key

In the setup of the scheme, Alice publishes her public key A = G x [a], where a is the private key chosen uniformly at random from [1, n-1].

respond

Alice computes r = v - a * c mod n and sends it to Bob.

verify

Verify A is a valid point on the curve and A x [h] is not the point at infinity. To verify V = G x [r] + A x [c].